# [FLINK-13498][kafka] abort transactions in parallel