@NicoK 2019-08-01
Thanks @pnowojski for the review. Following a discussion with @becketqin, I reverted the parts that used our own `ForkJoinPool` and changed the code to use Java's default/common pool. That greatly simplifies the code but I was initially also worried and hence created that version.